Update to Rails 4.2.5.1, address UJS issue
Along the way of upgrading to Rails 4.2.5.1 from 4.2.4, I found a test
that was failing with Capybara. That looked to be due to the version of
qt5 that I had installed. Rectified by rolling back to qt 4, and
reinstalling Capybara and the capybara-webkit driver.

From there, I found that how I was using UJS to fetch a feed wasn't
passing testing, so updated to use a :script type on the link_to remote
call and use a JS template. That allows me to remove the feed_entries.js
file altogether.

This is the direction I want to go anyhow so, :)